Refactor: separate Data and Domain according to our architecture
