+++ ivy-data/src/main/java/com/ivy/data/model/Account.kt
@@ -1,9 +1,9 @@
-package com.ivy.domain.model
+package com.ivy.data.model
 
 import androidx.compose.ui.graphics.Color
-import com.ivy.domain.model.primitive.AssetCode
-import com.ivy.domain.model.primitive.IconId
-import com.ivy.domain.model.primitive.NotBlankTrimmedString
+import com.ivy.data.model.primitive.AssetCode
+import com.ivy.data.model.primitive.IconId
+import com.ivy.data.model.primitive.NotBlankTrimmedString
 import java.util.UUID
 
 data class Account(
