+++ ivy-data/src/main/java/com/ivy/data/model/primitive/PositiveDouble.kt
@@ -1,8 +1,8 @@
-package com.ivy.domain.model.primitive
+package com.ivy.data.model.primitive
 
 import arrow.core.raise.Raise
 import arrow.core.raise.ensure
-import com.ivy.domain.exact.Exact
+import com.ivy.base.exact.Exact
 
 @JvmInline
 value class PositiveDouble private constructor(val value: Double) {
