+++ ivy-domain/build.gradle.kts
@@ -10,6 +10,5 @@ dependencies {
     implementation(projects.ivyBase)
     implementation(projects.ivyData)
 
-    implementation(libs.datastore)
-    implementation(libs.bundles.ktor)
+    implementation(libs.datastore) // TODO: Will be removed
 }
