+++ ivy-domain/src/test/java/com/ivy/domain/model/primitive/NotBlankTrimmedStringTest.kt
@@ -1,5 +1,6 @@
 package com.ivy.domain.model.primitive
 
+import com.ivy.data.model.primitive.NotBlankTrimmedString
 import io.kotest.assertions.arrow.core.shouldBeLeft
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.booleans.shouldBeTrue
