+++ ivy-domain/src/test/java/com/ivy/domain/model/primitive/PositiveDoubleTest.kt
@@ -1,9 +1,9 @@
 package com.ivy.domain.model.primitive
 
+import com.ivy.data.model.primitive.PositiveDouble
 import io.kotest.assertions.arrow.core.shouldBeLeft
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.booleans.shouldBeTrue
-import io.kotest.matchers.shouldBe
 
 class PositiveDoubleTest : FreeSpec({
     "fails for" - {
