+++ screen-piechart/src/main/java/com/ivy/piechart/PieChart.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.viewinterop.AndroidView
-import com.ivy.base.util.toEpochMilli
+import com.ivy.base.legacy.toEpochMilli
 import com.ivy.legacy.datamodel.Category
 import com.ivy.design.l0_system.UI
 import com.ivy.legacy.utils.drawColoredShadow
