+++ temp-legacy-code/src/main/java/com/ivy/legacy/IntervalTypeExt.kt
@@ -1,6 +1,6 @@
 package com.ivy.legacy
 
-import com.ivy.base.util.stringRes
+import com.ivy.base.legacy.stringRes
 import com.ivy.data.model.IntervalType
 import com.ivy.resources.R
 import java.time.LocalDateTime
