+++ temp-legacy-code/src/main/java/com/ivy/legacy/IvyComposeApp.kt
@@ -12,8 +12,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalView
-import com.ivy.base.model.Theme
-import com.ivy.base.util.appContext
+import com.ivy.base.legacy.Theme
+import com.ivy.base.legacy.appContext
 import com.ivy.domain.RootScreen
 import com.ivy.design.IvyContext
 import com.ivy.design.api.IvyDesign
