+++ temp-legacy-code/src/main/java/com/ivy/legacy/LoanExt.kt
@@ -1,6 +1,6 @@
 package com.ivy.legacy
 
-import com.ivy.base.util.stringRes
+import com.ivy.base.legacy.stringRes
 import com.ivy.legacy.datamodel.Loan
 import com.ivy.data.model.LoanType
 import com.ivy.resources.R
