+++ temp-legacy-code/src/main/java/com/ivy/legacy/data/model/Month.kt
@@ -1,7 +1,7 @@
 package com.ivy.legacy.data.model
 
 import androidx.compose.runtime.Immutable
-import com.ivy.base.util.stringRes
+import com.ivy.base.legacy.stringRes
 import com.ivy.legacy.IvyWalletCtx
 import com.ivy.legacy.utils.dateNowUTC
 import com.ivy.resources.R
