+++ temp-legacy-code/src/main/java/com/ivy/legacy/datamodel/Settings.kt
@@ -1,7 +1,7 @@
 package com.ivy.legacy.datamodel
 
 import androidx.compose.runtime.Immutable
-import com.ivy.base.model.Theme
+import com.ivy.base.legacy.Theme
 import com.ivy.data.db.entity.SettingsEntity
 import java.math.BigDecimal
 import java.util.UUID
