+++ temp-legacy-code/src/main/java/com/ivy/legacy/notification/TransactionReminderWorker.kt
@@ -6,7 +6,7 @@ import androidx.core.app.NotificationCompat
 import androidx.hilt.work.HiltWorker
 import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
-import com.ivy.base.util.stringRes
+import com.ivy.base.legacy.stringRes
 import com.ivy.domain.AppStarter
 import com.ivy.legacy.data.SharedPrefs
 import com.ivy.legacy.utils.atEndOfDay
